+++ src/splits_to_tree.cpp
@@ -5,7 +5,7 @@ using namespace Rcpp;
 #include "../inst/include/TreeTools/renumber_tree.h"
 using namespace TreeTools;
 
-TREETOOLS_SPLITLIST_INIT;
+TREETOOLS_SPLITLIST_INIT
 
 inline void insertion_sort_by_largest(int16* arr, const int16 arr_len,
                                        const int16* sort_by) {